split main.js
