+++ dashboard/apps/prototype/static/prototype/js/main.js
@@ -1,63 +1,11 @@
-import 'whatwg-fetch';
-import Cookies from 'js-cookie';
-import Plotly from 'plotly.js/lib/core';
-import * as bar from 'plotly.js/lib/bar';
-Plotly.register([bar]);
-
 import $ from 'jquery';
 import select2 from 'select2';
-import URI from 'urijs';
-import _ from 'lodash';
-import moment from 'moment';
+import Plotly from './custom-plotly';
+import {getProjectId, loadProjectPage, parseProjectFinancials, getProjectData} from './project';
 
 require('select2/dist/css/select2.min.css');
 require('../styles/gov-uk-elements.css');
 require('../styles/main.css');
-
-
-/**
- * send a POST request to the backend to retrieve project profile
- */
-function getProjectData(id) {
-  return fetch('/project.json', {
-    credentials: 'same-origin',
-    method: 'POST',
-    headers: {
-      'X-CSRFToken': Cookies.get('csrftoken'),
-      'Accept': 'application/json',
-      'Content-Type': 'application/json'
-    },
-    body: JSON.stringify({projectid: id})
-  }).then(response => response.json());
-}
-
-/**
- * parse the financial infomation about the project
- */
-function parseProjectFinancials(financial) {
-  let [months, costs] = _(financial).toPairs().sort().unzip().value();
-  months = months.map(m => moment(m, 'YYYY-MM').format('MMM YY'));
-
-  const _mapFloat = key => costs.map(c => parseFloat(c[key]));
-  const contractorCosts = _mapFloat('contractor');
-  const civilServantCosts = _mapFloat('non-contractor');
-  const additionalCosts = _mapFloat('additional');
-  const budget = _mapFloat('budget');
-  
-  const totalCosts = _.zip(contractorCosts, civilServantCosts, additionalCosts)
-                      .map(([x, y, a]) => x + y + a);
-  const totalCostsCumulative = [];
-  totalCosts.reduce((x, y, i) => totalCostsCumulative[i] = x + y, 0);
-
-  return {
-    months,
-    budget,
-    civilServantCosts,
-    contractorCosts,
-    additionalCosts,
-    totalCostsCumulative
-  };
-}
 
 /**
  * plot the graphs for a project
@@ -136,23 +84,6 @@ function plotProject(project, elem) {
 }
 
 
-/**
- * get projectId based on the query string
- */
-function getProjectId() {
-  return URI(window.location.href).query(true).projectid;
-}
-
-
-/**
- * load the page for the project based on id
- **/
-function loadProjectPage(id) {
-  const url = [location.protocol, '//', location.host, location.pathname].join('');
-  window.location.href = url + '?projectid=' + id;
-}
-
-  
 $(() => {
   // get the DOM element for the graph
   const elem = document.getElementById('fig-a');
